Initiate Neo4j-OGM session through OSGi components
Use classes OGM-Session-Component and OGM-SessionFactory (OSGi-based component, respectively factory) to initiate a Neo4j-OGM session. Due to class-loading issues, we decided to delegate the initialization of a Neoj4-OGM session factory to a different module, direct utilization of Neo4j-OGM Session-Factory gets avoided.

Important:
The class Generic-Service is abstract; there is an injecting here. Due to this aspect, you have to instruct the maven bundle plugin to inherit ds-annotations.